backendApi/requirements.txt

asgiref==3.8.1
certifi==2024.2.2

charset-normalizer==3.3.2
cloudinary==1.40.0


distlib==0.3.8
Django==5.0.6
django-cors-headers==4.3.1
django-filter==24.2
djangorestframework==3.15.1
filelock==3.14.0

gunicorn==22.0.0
idna==3.7


numpy==1.26.4
packaging==24.0

pillow==10.3.0
platformdirs==4.2.1
pygbag==0.9.1



pyperclip==1.8.2
pypng==0.20220715.0

python-dateutil==2.9.0.post0
pytz==2024.1
qrcode==7.4.2
requests==2.31.0
six==1.16.0
sqlparse==0.5.0
tk==0.1.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
virtualenv==20.26.2
whitenoise==6.6.0
